perf(prims-ts): match high-throughput NVFP4 MoE performance
Align the end-to-end tensor adaptation, memory-resource, and kernel launch paths used by high-throughput NVFP4 MoE workloads.

+++ flashinfer/fused_moe/backends/prims_ts/fp4_op.py
@@ -72,7 +72,7 @@ def _resolve_routing_inputs(
         return (
             routing_logits,
             torch.empty(0, dtype=torch.int32, device=hidden_states.device),
-            torch.empty(0, dtype=routing_logits.dtype, device=hidden_states.device),
+            torch.empty(0, dtype=torch.bfloat16, device=hidden_states.device),
         )
 
     if topk_ids is None:
